Moved util tests
